feat: rename tailormade to boxing
